+++ src/ape/mod.rs
@@ -27,7 +27,7 @@ use std::io::{Read, Seek};
 cfg_if::cfg_if! {
 	if #[cfg(feature = "ape")] {
 		pub(crate) mod tag;
-		pub use tag::ape_tag::ApeTag;
+		pub use tag::ApeTag;
 		pub use tag::item::ApeItem;
 
 		pub use crate::types::picture::APE_PICTURE_TYPES;